Verification Warning & Freeze Authority update (#446)
This PR is a response to Meow's request for community support updating Jupiter's documentation regarding Freeze Authority.

Initially "Freeze Authority & Verification Warning" were a single section. For the purposes of clarity, I have split them into two separate sections.

It should also be noted that for the images for both the Verification Warning and Freeze Authority should be updated once the UI is updated.